Implemented api run-simulation endpoint in domain, persistence and api layers

+++ api/src/main/kotlin/nl/sogyo/modelr/models/ReactorSettingsDTO.kt
@@ -16,7 +16,7 @@ data class ReactorSettingsDTO(
     @JsonProperty
     var impellerType: String, //g gx-1 h-1
     @JsonProperty
-    var numberOfImpellers: Double, //-
+    var numberOfImpellers: Int, //-
     @JsonProperty
     var agitatorSpeed: Double,
 )
